Singly LinkedList and CircularLinkedList
There were implemented as SinglyLinkedLIst as well as CircularLinkedList.
The benefit of CircularLinkedList is that we have the variable last , which refers to the last node of our list. Due to this we can achieve the minimal time when we are gonna insert a data in the last. Time in this way is O(1)